- removal of LanguagesListActivity.kt and its related codes - renames some classes
